+++ .erb/scripts/delete-source-maps.js
@@ -3,6 +3,8 @@ import rimraf from 'rimraf';
 import webpackPaths from '../configs/webpack.paths';
 
 export default function deleteSourceMaps() {
-  rimraf.sync(path.join(webpackPaths.distMainPath, '*.js.map'));
-  rimraf.sync(path.join(webpackPaths.distRendererPath, '*.js.map'));
+  if (fs.existsSync(webpackPaths.distMainPath))
+    rimraf.sync(path.join(webpackPaths.distMainPath, '*.js.map'));
+  if (fs.existsSync(webpackPaths.distRendererPath))
+    rimraf.sync(path.join(webpackPaths.distRendererPath, '*.js.map'));
 }
